\name{GMDH-package}
\alias{GMDH-package}
\alias{GMDH-package}
\docType{package}
\title{
Predicting and Forecasting Time Series via GMDH-Type Neural Network Algorithms
}
\description{
Package GMDH includes a function for predicting and forecasting a univariate time series by using GMDH-type neural network, and a dataset for implementation.

Dag, O., Yozgatligil, C. (2015). GMDH: An R Package for Predicting and Forecasting Time Series via GMDH-Type Neural Network Algorithms. To be submitted.

}
\details{
\tabular{ll}{
Package: \tab GMDH\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-25\cr
License: \tab GPL (>=2)\cr
}
}