\name{GMDH-package}
\alias{GMDH-package}
\alias{GMDH-package}
\docType{package}
\title{
Predicting and Forecasting Time Series via GMDH-Type Neural Network Algorithms
}
\description{
Package GMDH includes a function for predicting and forecasting a univariate time series by using GMDH-type neural network, and a dataset for implementation.
}
\details{
\tabular{ll}{
Package: \tab GMDH\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2015-07-20\cr
License: \tab GPL (>=2)\cr
}
}