\name{GMDH-package}
\alias{GMDH-package}
\alias{GMDH-package}
\docType{package}
\title{
Short Term Forecasting via GMDH-Type Neural Network Algorithms
}
\description{
Package GMDH includes a function for short term forecasting of a univariate time series by using GMDH-type neural network algorithms, and a dataset for implementation.
}
\details{
\tabular{ll}{
Package: \tab GMDH\cr
Type: \tab Package\cr
Version: \tab 1.6\cr
Date: \tab 2016-09-20\cr
License: \tab GPL (>=2)\cr
}
}